backend/requirements/base.txt
# Core Django
Django==5.0.1
django-environ==0.11.2
psycopg2-binary==2.9.9
psycopg[binary]==3.1.16

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0

# Database & Caching
redis==5.0.1
django-redis==5.4.0
pgvector==0.2.4

# Celery
celery==5.3.4
celery[redis]==5.3.4
django-celery-beat==2.5.0
flower==2.0.1

# Security & Encryption
django-cryptography==1.1
django-guardian==2.4.0
pyotp==2.9.0
qrcode==7.4.2

# Authentication & OAuth
oauthlib==3.2.2
requests-oauthlib==1.3.1

# AI & ML
google-generativeai==0.3.2
langchain==0.1.0
chromadb==0.4.22
openai==1.7.2

# Document Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pytesseract==0.3.10

# HTTP & API Clients
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-slugify==8.0.1
shortuuid==1.0.11

# Email
sendgrid==6.11.0

# SMS
twilio==8.11.1

# Validation
jsonschema==4.20.0
pydantic==2.5.3

# Monitoring & Logging
sentry-sdk==1.39.2
python-json-logger==2.0.7

# Rate Limiting
django-ratelimit==4.1.0

# CORS
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Channels (WebSocket)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Storage
django-storages==1.14.2
boto3==1.34.19

# Template Engine
Jinja2==3.1.3

# Task Scheduling
django-celery-results==2.5.1

# Utils
python-magic==0.4.27
user-agents==2.2.0
